+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Version [2.1.1] - 2024-02-01
+
+### Changed
+
+- Zurg: Download release version parsing using GitHub release tags
+
+### Fixed
+
+- Healthcheck: Fixed healthcheck for rclone serve NFS  
+
+
 ## Version [2.1.0] - 2024-01-29
 
 ### Added